Fix: toggling between higlight-comment/highlight annotations
- Fixes toggling when the last comment is deleted on a highlight-comment annotation and the user doesn't have permissions to delete the ENTIRE annotation
- Prevent error from being thrown when deleting plain highlights

+++ src/AnnotationThread.js
@@ -259,6 +259,7 @@ class AnnotationThread extends EventEmitter {
             this.saveAnnotation(TYPES.highlight);
         }
         this.type = TYPES.highlight_comment;
+        this.state = STATES.pending;
         this.renderAnnotationPopover();
     }
 
@@ -298,11 +299,20 @@ class AnnotationThread extends EventEmitter {
         // Delete matching comment from annotation
         this.comments = this.comments.filter(({ id }) => id !== annotationIDToRemove);
 
-        if (this.type === TYPES.highlight && !this.canDelete) {
+        if (this.type === TYPES.highlight_comment && this.comments.length < 0) {
+            this.type = TYPES.highlight;
+
             // If the user doesn't have permission to delete the entire highlight
             // annotation, display the annotation as a plain highlight
-            this.cancelFirstComment();
-        } else if (this.type === TYPES.highlight || this.type === TYPES.draw || this.comments.length <= 0) {
+            if (!this.canDelete) {
+                this.cancelFirstComment();
+            }
+        }
+
+        if (
+            this.canDelete &&
+            (this.type === TYPES.highlight || this.type === TYPES.draw || this.comments.length <= 0)
+        ) {
             // If this annotation was the last one in the thread, destroy the thread
             this.destroy();
             this.threadID = null;
@@ -327,7 +337,7 @@ class AnnotationThread extends EventEmitter {
             .then(() => {
                 // Ensures that blank highlight comment is also deleted when removing
                 // the last comment on a highlight
-                if (this.type === TYPES.highlight && this.canDelete) {
+                if (this.threadID && this.type === TYPES.highlight && this.canDelete) {
                     this.api.delete(this.id).then(() => {
                         // Broadcast thread cleanup if needed
                         if (!this.threadID) {
